backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0
# Email notifications
fastapi-mail==1.4.1
# Calendar integrations
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
# Payment integration
stripe==7.8.0
# Video conferencing
requests==2.31.0
# Recurring patterns
python-dateutil==2.8.2
# SSO Authentication
python-jose[cryptography]==3.3.0
httpx==0.25.2
# ICS Calendar integration
icalendar==5.0.7
